+++ src/mmv/mmvskia/mmv_context.py
@@ -32,71 +32,20 @@ import sys
 import os
 
 
-class MMVSkiaPath:
-    def __init__(self, mmv_main, depth = LOG_NO_DEPTH):
-        debug_prefix = "[MMVSkiaPath.__init__]"
-        ndepth = depth + LOG_NEXT_DEPTH
-        self.mmv_main = mmv_main
-
-        # The root directory we'll be refering to, it's the folder where __init__.py is located
-        # We need to get the mmv_main's interface class position as the "package mmv" is the interface itself
-        self.MMV_SKIA_ROOT = self.mmv_main.mmvskia_interface.MMV_SKIA_ROOT
-
-        # Path separator
-        sep = os.path.sep
-
-        # # Externals directory (ffmpeg binaries mainly on Windows)
-        
-        # Define it
-        self.externals_dir = f"{self.MMV_SKIA_ROOT}{sep}externals"
-        logging.info(f"{depth}{debug_prefix} Externals folder path is: [{self.externals_dir}]")
-            
-        # Make the directory if it doesn't exist
-        logging.info(f"{depth}{debug_prefix} Make directory if doesn't exist Externals Directory")
-        self.mmv_main.utils.mkdir_dne(self.externals_dir, depth = ndepth)
-
-        # Append to PATH the 
-        logging.info(f"{depth}{debug_prefix} Appending Externals Directory to system PATH environment variable")
-        sys.path.append(self.externals_dir)
-
-        # # Data directory (configurations)
-
-        self.data_dir = f"{self.MMV_SKIA_ROOT}{sep}data"
-        logging.info(f"{depth}{debug_prefix} Data folder path is: [{self.data_dir}]")
-
-        # # Error assertion the directories
-
-        self.error_assertion(depth = ndepth)
-
-    # Make sure the directories and files we'll need exist (required ones)
-    def error_assertion(self, depth = LOG_NO_DEPTH):
-        debug_prefix = "[MMVSkiaPath.error_assertion]"
-        ndepth = depth + LOG_NEXT_DEPTH
-
-        # # Directories
-        
-        self.mmv_main.utils.assert_dir(self.data_dir, depth = ndepth)
-        self.mmv_main.utils.assert_dir(self.externals_dir, depth = ndepth)
-        
-
+# Also see __init__.py with MMVSkiaInterface
+# that class also changes some configurations here, this is motly
+# shared stuff across files and it's a bit hard to keep track of them all
 class MMVContext:
     def __init__(self, mmv_main, depth = LOG_NO_DEPTH) -> None:
         debug_prefix = "[MMVContext.__init__]"
         ndepth = depth + LOG_NEXT_DEPTH
         self.mmv_main = mmv_main
 
-        # Create classes
-        logging.info(f"{depth}{debug_prefix} Creating MMVSkiaPath() class")
-        self.paths = MMVSkiaPath(self.mmv_main, ndepth)
-
         # Files, info
         self.output_video = None
-        self.input_file = None
+        self.input_audio_file = None
         self.input_midi = None
         self.duration = None
-
-        # External dependencies directory
-        self.data_dir = self.paths.MMV_SKIA_ROOT + "/data"
 
         # Video specs
         self.width = 1280