+++ src/mmv/mmvskia/mmv_core.py
@@ -38,11 +38,11 @@ import os
 
 
 class MMVSkiaCore:
-    def __init__(self, mmv_main, depth = LOG_NO_DEPTH) -> None:
+    def __init__(self, mmvskia_main, depth = LOG_NO_DEPTH) -> None:
         debug_prefix = "[MMVSkiaCore.__init__]"
         ndepth = depth + LOG_NEXT_DEPTH
-        self.mmv_main = mmv_main
-        self.prelude = self.mmv_main.prelude
+        self.mmvskia_main = mmvskia_main
+        self.prelude = self.mmvskia_main.prelude
         self.preludec = self.prelude["mmvcore"]
 
         # Log creation
@@ -59,7 +59,7 @@ class MMVSkiaCore:
 
         # # Save info so we can utilize on post processing or somewhere else
 
-        last_session_info_file = self.mmv_main.mmvskia_interface.top_level_interace.last_session_info_file
+        last_session_info_file = self.mmvskia_main.mmvskia_interface.top_level_interace.last_session_info_file
         logging.info(f"{depth}{debug_prefix} Saving partial session info to last_session_info file at [{last_session_info_file}]")
 
         # Quit if code flow says so
@@ -73,19 +73,79 @@ class MMVSkiaCore:
 
         # Read the audio and start FFmpeg pipe
         logging.info(f"{depth}{debug_prefix} Read audio file")
-        self.mmv_main.audio.read(path = self.mmv_main.context.input_file, depth = ndepth)
+        self.mmvskia_main.audio.read(path = self.mmvskia_main.context.input_audio_file, depth = ndepth)
         
+        # How many steps is the audio duration times the frames per second
+        self.mmvskia_main.context.total_steps = int(self.mmvskia_main.audio.duration * self.mmvskia_main.context.fps)
+        logging.info(f"{depth}{debug_prefix} Total steps: {self.mmvskia_main.context.total_steps}")
+
+        # Update info that might have been changed by the user
+        logging.info(f"{depth}{debug_prefix} Update Context bases")
+        self.mmvskia_main.context.update_biases()
+
         # Create the pipe write thread
         if not ONLY_PROCESS_AUDIO:
            
+            # Set pixel format according to the OS
+            if self.mmvskia_main.context.ffmpeg_pixel_format == "auto":
+                logging.info(f"{depth}{debug_prefix} Pixel format is [auto], getting right one based on the OS..")
+
+                # Windows
+                if self.mmvskia_main.utils.os == "windows":
+                    logging.info(f"{depth}{debug_prefix} Pixel format set to [bgra] because Windows OS")
+                    pixel_format = "bgra"
+
+                # Linux
+                elif self.mmvskia_main.utils.os == "linux":
+                    logging.info(f"{depth}{debug_prefix} Pixel format set to [rgba] because GNU/Linux OS")
+                    pixel_format = "rgba"
+                
+                # MacOS
+                elif self.mmvskia_main.utils.os == "macos":
+                    logging.info(f"{depth}{debug_prefix} Pixel format set to [rgba] because Darwin / MacOS")
+                    pixel_format = "rgba"
+
+                else: # Not configured, found?
+                    raise RuntimeError(f"Pixel format \"auto\" not found for OS: [{self.mmvskia_main.utils.os}]")
+            else:
+                pixel_format = self.mmvskia_main.context.ffmpeg_pixel_format
+
             # Start video pipe
             logging.info(f"{depth}{debug_prefix} Starting FFmpeg Pipe")
-            self.mmv_main.ffmpeg.pipe_one_time(self.mmv_main.context.output_video)
+            self.mmvskia_main.ffmpeg.pipe_images_to_video(
 
+                # Search for a FFmpeg binary
+                ffmpeg_binary_path = self.mmvskia_main.utils.get_executable_with_name(
+                    "ffmpeg",
+                    extra_paths = self.mmvskia_main.mmvskia_interface.top_level_interace.externals_dir,
+                    depth = ndepth    
+                ),
+
+                # Dump MMVContext configuration
+                width = self.mmvskia_main.context.width,
+                height = self.mmvskia_main.context.height,
+                input_audio_file = self.mmvskia_main.context.input_audio_file,
+                output_video = self.mmvskia_main.context.output_video,
+                pix_fmt = pixel_format,
+                framerate = self.mmvskia_main.context.fps,
+                preset = self.mmvskia_main.context.x264_preset,
+                hwaccel = self.mmvskia_main.context.ffmpeg_hwaccel,
+                opencl = self.mmvskia_main.context.x264_use_opencl,
+                dumb_player = self.mmvskia_main.context.ffmpeg_dumb_player,
+                crf = self.mmvskia_main.context.x264_crf,
+                depth = ndepth,
+            )
+
+            # Create pipe writer thread
             logging.info(f"{depth}{debug_prefix} Creating pipe writer thread")
             self.pipe_writer_loop_thread = threading.Thread(
-                target = self.mmv_main.ffmpeg.pipe_writer_loop,
-                args = (self.mmv_main.audio.duration,),
+                target = self.mmvskia_main.ffmpeg.pipe_writer_loop,
+                args = (
+                    self.mmvskia_main.audio.duration,
+                    self.mmvskia_main.context.fps,
+                    self.mmvskia_main.context.total_steps,
+                    self.mmvskia_main.context.max_images_on_pipe_buffer
+                ),
                 daemon = True,
             )
 
@@ -95,19 +155,11 @@ class MMVSkiaCore:
 
             # Init Skia
             logging.info(f"{depth}{debug_prefix} Init Skia")
-            self.mmv_main.skia.init(
-                width = self.mmv_main.context.width,
-                height = self.mmv_main.context.height,
-                render_backend = self.mmv_main.context.skia_render_backend,
+            self.mmvskia_main.skia.init(
+                width = self.mmvskia_main.context.width,
+                height = self.mmvskia_main.context.height,
+                render_backend = self.mmvskia_main.context.skia_render_backend,
             )
-
-        # How many steps is the audio duration times the frames per second
-        self.mmv_main.context.total_steps = int(self.mmv_main.audio.duration * self.mmv_main.context.fps)
-        logging.info(f"{depth}{debug_prefix} Total steps: {self.mmv_main.context.total_steps}")
-
-        # Update info that might have been changed by the user
-        logging.info(f"{depth}{debug_prefix} Update Context bases")
-        self.mmv_main.context.update_biases()
 
         # What to log and what not to
         LOG_STEP = self.preludec["run"]["log_step"]
@@ -128,15 +180,15 @@ class MMVSkiaCore:
         # # Last session info
 
         # Reset last session info file
-        self.mmv_main.utils.reset_file(last_session_info_file)
+        self.mmvskia_main.utils.reset_file(last_session_info_file)
         
         # Dump to toml file
-        self.mmv_main.utils.dump_toml(
+        self.mmvskia_main.utils.dump_toml(
             data = {
-                "output_video": self.mmv_main.context.output_video,
-                "frame_count": self.mmv_main.context.total_steps,
-                "width": self.mmv_main.context.width,
-                "height": self.mmv_main.context.height,
+                "output_video": self.mmvskia_main.context.output_video,
+                "frame_count": self.mmvskia_main.context.total_steps,
+                "width": self.mmvskia_main.context.width,
+                "height": self.mmvskia_main.context.height,
             },
             path = last_session_info_file
         )
@@ -144,10 +196,10 @@ class MMVSkiaCore:
         # # Main routine
 
         logging.info(f"{depth}{debug_prefix} Start main routine")
-        logging.info(f"{depth}{debug_prefix} Video will be saved in [{self.mmv_main.context.output_video}]")
+        logging.info(f"{depth}{debug_prefix} Video will be saved in [{self.mmvskia_main.context.output_video}]")
 
         # Iterate over all steps
-        for step in range(0, self.mmv_main.context.total_steps):
+        for step in range(0, self.mmvskia_main.context.total_steps):
 
             # Log current step, next iteration
             if LOG_STEP:
@@ -160,37 +212,37 @@ class MMVSkiaCore:
             # # # [ Slice the audio ] # # #
 
             # Add the offset audio step (because interpolation isn't instant for smoothness)
-            self.this_step = step + self.mmv_main.context.offset_audio_before_in_many_steps
+            self.this_step = step + self.mmvskia_main.context.offset_audio_before_in_many_steps
 
             # If this step is out of bounds because the offset, set it to its max value
-            if self.this_step >= self.mmv_main.context.total_steps - 1:
-                self.this_step = self.mmv_main.context.total_steps - 1
+            if self.this_step >= self.mmvskia_main.context.total_steps - 1:
+                self.this_step = self.mmvskia_main.context.total_steps - 1
 
             # Log offset step
             if LOG_OFFSETTED_STEP:
-                logging.debug(f"{depth}{debug_prefix} Offsetted step by [{self.mmv_main.context.offset_audio_before_in_many_steps}] is [{self.this_step}]")
+                logging.debug(f"{depth}{debug_prefix} Offsetted step by [{self.mmvskia_main.context.offset_audio_before_in_many_steps}] is [{self.this_step}]")
 
             # The current time in seconds we're going to slice the audio based on its sample rate
             # If we offset to the opposite way, the starting point can be negative hence the max function.
-            current_time = max((1/self.mmv_main.context.fps) * self.this_step, 0)
+            current_time = max((1/self.mmvskia_main.context.fps) * self.this_step, 0)
 
             # Current time we're processing
-            self.mmv_main.context.current_time = (1/self.mmv_main.context.fps) * self.this_step
+            self.mmvskia_main.context.current_time = (1/self.mmvskia_main.context.fps) * self.this_step
 
             # The current time in sample count to slice the audio
-            this_time_in_samples = int(current_time * self.mmv_main.audio.sample_rate)
+            this_time_in_samples = int(current_time * self.mmvskia_main.audio.sample_rate)
 
             # The slice starts at the this_time_in_samples and end the cut here
-            until = int(this_time_in_samples + self.mmv_main.context.batch_size)
+            until = int(this_time_in_samples + self.mmvskia_main.context.batch_size)
 
             # Slice the audio
-            self.mmv_main.audio_processing.slice_audio(
-                stereo_data = self.mmv_main.audio.stereo_data,
-                mono_data = self.mmv_main.audio.mono_data,
-                sample_rate = self.mmv_main.audio.sample_rate,
+            self.mmvskia_main.audio_processing.slice_audio(
+                stereo_data = self.mmvskia_main.audio.stereo_data,
+                mono_data = self.mmvskia_main.audio.mono_data,
+                sample_rate = self.mmvskia_main.audio.sample_rate,
                 start_cut = this_time_in_samples,
                 end_cut = until,
-                batch_size = self.mmv_main.context.batch_size
+                batch_size = self.mmvskia_main.context.batch_size
             )
 
             # # # [ Calculate the FFTs ] # # #
@@ -199,12 +251,12 @@ class MMVSkiaCore:
             frequencies_list = []
 
             # For each sliced channel data we have, process that into the FFTs list
-            for channel_data in self.mmv_main.audio_processing.audio_slice:
+            for channel_data in self.mmvskia_main.audio_processing.audio_slice:
                
                 # Process this audio sample
-                fft, frequencies = self.mmv_main.audio_processing.process(
+                fft, frequencies = self.mmvskia_main.audio_processing.process(
                     data = channel_data,
-                    original_sample_rate = self.mmv_main.audio.sample_rate,
+                    original_sample_rate = self.mmvskia_main.audio.sample_rate,
                 )
 
                 # Add to the lists
@@ -213,7 +265,7 @@ class MMVSkiaCore:
 
             # We can access this dictionary from anyone for this step audio information
             self.modulators = {
-                "average_value": self.mmv_main.audio_processing.average_value * self.mmv_main.context.audio_amplitude_multiplier,
+                "average_value": self.mmvskia_main.audio_processing.average_value * self.mmvskia_main.context.audio_amplitude_multiplier,
                 "fft": fft_list,
                 "frequencies": frequencies_list,
             }
@@ -234,42 +286,42 @@ class MMVSkiaCore:
                 # Reset skia canvas
                 if LOG_NEXT_STEPS:
                     logging.debug(f"{depth}{debug_prefix} Reset skia canvas")
-                self.mmv_main.skia.reset_canvas()
+                self.mmvskia_main.skia.reset_canvas()
 
                 # Process next animation with audio info and the step count to process on
                 if LOG_NEXT_STEPS:
                     logging.debug(f"{depth}{debug_prefix} Call MMVSkiaAnimation.next()")
-                self.mmv_main.mmv_animation.next()
+                self.mmvskia_main.mmv_animation.next()
 
                 # Next image to pipe
                 if LOG_NEXT_STEPS:
                     logging.debug(f"{depth}{debug_prefix} Get next image from canvas array")
-                next_image = self.mmv_main.skia.canvas_array()
+                next_image = self.mmvskia_main.skia.canvas_array()
 
                 # Save current canvas's Frame to the final video, the pipe writer thread will actually pipe it
                 if LOG_NEXT_STEPS:
                     logging.debug(f"{depth}{debug_prefix} Write image to FFmpeg pipe index [{global_frame_index}]")
-                self.mmv_main.ffmpeg.write_to_pipe(global_frame_index, next_image)
+                self.mmvskia_main.ffmpeg.write_to_pipe(global_frame_index, next_image)
             
             else:  # QOL print what is happening
-                print(f"\rOnly process audio [{global_frame_index} / {self.mmv_main.context.total_steps}", end="")
+                print(f"\rOnly process audio [{global_frame_index} / {self.mmvskia_main.context.total_steps}", end="")
 
         # End pipe, no pipe to close if we're only processing audio
         if ONLY_PROCESS_AUDIO:
-            self.mmv_main.skia.terminate_glfw()
+            self.mmvskia_main.skia.terminate_glfw()
         else:
             logging.info(f"{depth}{debug_prefix} Call to close pipe, let it wait until it's done")
-            self.mmv_main.ffmpeg.close_pipe()
+            self.mmvskia_main.ffmpeg.close_pipe()
 
         # Update the TOML with the new data
         if WRITE_AUDIO_AMPLITUDE_VALUES_TO_LAST_SESSION_INFO:
 
             # Load and change the key we're interested in
-            previous_session_info_data = self.mmv_main.utils.load_toml(path = last_session_info_file)
+            previous_session_info_data = self.mmvskia_main.utils.load_toml(path = last_session_info_file)
             previous_session_info_data["audio_amplitudes"] = recorded_audio_amplitudes
 
             # Dump to toml file
-            self.mmv_main.utils.dump_toml(
+            self.mmvskia_main.utils.dump_toml(
                 data = previous_session_info_data,
                 path = last_session_info_file,
             )