+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Overhaul shave some extra 10% speed on rendering times
 - Remove some deprecated code
 - Add logging timings on MMVImage
+- Add customizable encoding settings
+- Failsafe for reading audio
+- Use raw FFmpeg binary for piping videos into (shave ffmpeg-python dependency)
 
 ## [2.5.0] :: Unite :: (2020-12-25)
 - Merge both MMVSkia and MMVShader into a "common interface package", this allows: